tests
adds a CMakeLists.txt to prepare the testing framework, google test
and to build the binary to run the tests.

adds test_ext_sort.cpp which generates random unsigned integers and
tests the external merge sort algorithm for correctness. For this
commit, it contains a test for a small 800B file.