surplus data and merge step
check if we have more data to read and create an additional
chunk to contain that data.

for the merge step we manage several input buffers and one output
buffer.
We keep filling these buffers from the respective chunks and filling
the output buffer from the min-heap, flushing the output buffer to disk
when full.
Output file should contain the sorted data at the end.